refactor: rename env name and update select components
