+++ src/components/molecules/select/index.ts
@@ -1,3 +1,4 @@
+export { BaseSelect } from "@components/molecules/select/base";
 export { SelectCreatable } from "@components/molecules/select/creatable";
 export { IconLabel } from "@components/molecules/select/iconLabel";
 export { Select } from "@components/molecules/select/select";
