+++ src/vite-env.d.ts
@@ -3,7 +3,7 @@
 // / <reference types="vite-plugin-svgr/client" />
 
 interface ImportMetaEnv {
-	DISPLAY_COMBOX_TRIGGER: any;
+	readonly DISPLAY_COMBOX_IN_TRIGGERS_FORM: boolean;
 	readonly VITE_NODE_ENV: "development" | "production";
 	readonly VITE_AUTH_ENABLED: string;
 	readonly VITE_DESCOPE_PROJECT_ID: string;
