+++ vite.config.ts
@@ -43,7 +43,7 @@ export default defineConfig({
 		"import.meta.env.VITE_HOST_URL": JSON.stringify(process.env.VITE_HOST_URL),
 		"import.meta.env.DISPLAY_DISCORD_INTEGRATION": process.env.DISPLAY_DISCORD_INTEGRATION,
 		"import.meta.env.DISPLAY_SLACK_SOCKET_INTEGRATION": process.env.DISPLAY_SLACK_SOCKET_INTEGRATION,
-		"import.meta.env.DISPLAY_COMBOX_TRIGGER": process.env.DISPLAY_COMBOX_TRIGGER,
+		"import.meta.env.DISPLAY_COMBOX_IN_TRIGGERS_FORM": process.env.DISPLAY_COMBOX_IN_TRIGGERS_FORM,
 		"import.meta.env.SENTRY_DSN": JSON.stringify(process.env.SENTRY_DSN),
 		"import.meta.env.TESTS_JWT_AUTH_TOKEN": JSON.stringify(process.env.TESTS_JWT_AUTH_TOKEN),
 	},
